Split parameter store test to isolate boto3 lib requirements
